+++ lib/cineworld_uk.rb
@@ -5,6 +5,8 @@ require 'tzinfo/data'
 require_relative './cineworld_uk/version'
 
 require_relative './cineworld_uk/internal/api_response'
+require_relative './cineworld_uk/internal/parser/api/cinema_address'
+
 require_relative './cineworld_uk/internal/film_with_screenings_parser'
 require_relative './cineworld_uk/internal/screening_parser'
 require_relative './cineworld_uk/internal/title_sanitizer'